pyproject.toml
[tool.poetry]
name = "amusement-park-pipeline"
version = "0.1.0"
description = "A data engineering pipeline for amusement park queue times."
authors = ["Ronan Riboulet"]
packages = [
    { include = "src" },
    { include = "tools" },
    { include = "shared" }
]

[tool.poetry.dependencies]
# Define the Python version for this project
python = "^3.11"

# --- Project's main libraries here ---
pandas = "^2.1.3"
requests = "^2.31.0"
notebook = "^7.0.6"
ipykernel = "^6.27.1"
tenacity = "^9.1.2"
google-cloud-storage = "^3.6.0"
pydantic-settings = "^2.12.0"
google-cloud-bigquery = "^3.38.0"
ibis-framework = {extras = ["bigquery"], version = "^11.0.0"}
db-dtypes = "^1.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"